Add support for negation
This new feature adds support for a (simple) negation: meaning
the opposite of something. For example, whereas the word `bad`
has a valence of `-3`, the word `not bad` are in fact quite
positive.

This feature adds support for certain words, such as `not`,
`neither`, `nor`, and words ending in `n't`, to negate their
following word (if any).

Closes GH-1.